more delta time stuff

+++ SDL-Game/src/Player.h
@@ -13,7 +13,7 @@ public:
 	Player(Game* pGame);
 
 	void Start();
-	void Tick(FLOAT64 deltaTime);
+	void Tick(FLOAT32 deltaTime);
 	void Render(SDL_Renderer* pRenderer);
 
 private:
